--- Tiralabra_maven/src/main/java/com/mycompany/tiralabra_maven/Komento.java
@@ -1,6 +1,0 @@
-package com.mycompany.tiralabra_maven;
-
-
-public enum Komento {
-    MATRIISI, QUIT, DETERMINANTTI, KERRO, KOMENNOT, TULOSTALISTA, KAANNA;
-}
