+++ Tiralabra_maven/src/main/java/tiralabra_maven/App.java
@@ -1,4 +1,4 @@
-package com.mycompany.tiralabra_maven;
+package tiralabra_maven;
 
 /**
  * Hello world!
@@ -8,8 +8,9 @@ public class App
 {
     
     
-    public static void main( String[] args )
+    public static void main( String[] args ) throws Exception
     {
+        
         
         Kayttoliittyma kayttis = new Kayttoliittyma();
         kayttis.run();
