+++ Tiralabra_maven/src/main/java/tiralabra_maven/Komento.java
@@ -1,0 +1,9 @@
+package tiralabra_maven;
+
+/**
+ * Lista komennoista.
+ * @author risto
+ */
+public enum Komento {
+    MATRIISI, QUIT, DETERMINANTTI, KERRO, KOMENNOT, TULOSTALISTA, KAANNA, PLUS, MIINUS, SMONI, TULOSTA, TIEDOSTO;
+}
