+++ Tiralabra_maven/src/test/java/.LCKTiedostonkasittelijaTest.java~
@@ -1,0 +1,1 @@
+/home/risto/TiraLabra/Tiralabra_maven/src/test/java/TiedostonkasittelijaTest.java
