+++ Tiralabra_maven/src/test/java/KomennonKasittelijaTest.java
@@ -1,0 +1,39 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+import junit.framework.TestCase;
+
+/**
+ *
+ * @author risto
+ */
+public class KomennonKasittelijaTest extends TestCase {
+    
+    public KomennonKasittelijaTest(String testName) {
+        super(testName);
+    }
+    
+    @Override
+    protected void setUp() throws Exception {
+        super.setUp();
+    }
+    
+    @Override
+    protected void tearDown() throws Exception {
+        super.tearDown();
+    }
+
+    
+   
+    public void testtoimiikoTestaaminen() {
+        assertTrue(false);
+    }
+    
+    
+    
+    // TODO add test methods here. The name must begin with 'test'. For example:
+    // public void testHello() {}
+}
