+++ Tiralabra_maven/src/test/java/TiedostonkasittelijaTest.java
@@ -1,0 +1,31 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+import junit.framework.TestCase;
+
+/**
+ *
+ * @author risto
+ */
+public class TiedostonkasittelijaTest extends TestCase {
+    
+    public TiedostonkasittelijaTest(String testName) {
+        super(testName);
+    }
+    
+    @Override
+    protected void setUp() throws Exception {
+        super.setUp();
+    }
+    
+    @Override
+    protected void tearDown() throws Exception {
+        super.tearDown();
+    }
+
+    // TODO add test methods here. The name must begin with 'test'. For example:
+    // public void testHello() {}
+}
