+++ Tiralabra_maven/testi.txt
@@ -1,0 +1,2 @@
+1,2,3
+4,5,6
